requirements.txt
alembic==1.10.2
Flask==2.2.3
flask_migrate==4.0.4
flask_sqlalchemy==3.0.3
pandas==1.5.3
plotly==5.14.0
pydantic==1.10.7
PyDriller==2.4
python-dotenv==1.0.0
python_dateutil==2.8.2
psycopg2-binary==2.9.5